Early implementation of IPC
TLDR: It doesn't work

For some reason I get an exception on the client side that says the
connection was reset by peer. I will need to rework my server
implementation, but the client should be ready and it serves as somewhat
of a proof of concept on how the final IPC mechanism should work (hopefully).