IPC implementation working
Finally completed the implementation of the IPC protocol!! :)
Many files were modified, but it was all because the IPC protocol needed
a major restructure of almost everything.

The protocol works like so:
There's a server, that will be generating sudoku boards when requested,
with the specifications requested. Then, there's the client, that will
be requesting said board to the server. This is all done through a TCP
socket connection.

The protocol works like so:
1. The connection is established between the client and the server.
2. The client sends the properties of the wanted sudoku. The properties
are the size, represented through a 1 byte unsigned integer, that will
set the dimensions of the board (the board will always be squared), and
the difficulty, represented as another 1 byte integer, that will set how
many blank spaces will be left.
2. The server processes the properties requested by the client and
generates the according board.
3. The data is converted from a sudokuBoard class to a sbFile struct to
minimize the amount of data needed to be sent.
4. The server sends the board data to the client and the connection is
closed.
5. The client then uses the data in any way it wants (for now, we're
just printing the board).

Implementing all of this was a real pain as I had 0 previous experience
with networking, but thanks to Boost Asio's amazing tutorial and the
awesome sehe from
stackoverflow (https://stackoverflow.com/users/85371/sehe), I was able
to accomplish this :)